Getting there
Reading Main.ghx works now, and is very forgiving. It's line based, so everything on a line gets read individually and cleaned for ',' ' ' and  '.'. Works well. Still needs the ':' separator and ',' for multiple fileformats covered by the same script and fileformats are before ':' script is after. both with and without .ghx works  with the script reference.
Need to figure out dependency/listsystem. maybe a folder for it in each suite? idk. or maybe a seperate file format? (gets messy and busy though. Putting each script in a folder individually seems wasteful, unless I create a common resources folder/file system.