Working Read command
The scripting language '@ read' function now works. You can name it, if you dont it creates a randomly generated name.
Changed the files instance to be stored in Main(), this will simplify workflow in the future.
the language is more rigid now. needs spaces between all commands. Order is still somewhat free, it does need the @ to begin with.
Work on search function, this is whats next!

Revert and Write works fine still. But reverting to original still keeps the write button open.. Not sure if there is an easy way around this. This propably needs more testing and refinement.

Error reporting is needed for the scripting language. Need to figure out a way to handle this.